requirements.txt
streamlit>=1.31.0
openai>=1.12.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
beautifulsoup4>=4.12.0
scikit-learn>=1.4.0
